+++ ui-new/components/challenges/ChallengeCard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ChallengeType } from "@/store/challengeStore";
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/buttons";
+import { Button } from "@/components/ui/button";
 import { Trophy, Flag, FileText } from 'lucide-react';
 
 interface ChallengeCardProps {
