+++ ui-new/components/theme-toggle.tsx
@@ -4,7 +4,7 @@ import * as React from "react"
 import { Moon, Sun } from 'lucide-react'
 import { useTheme } from "next-themes"
 
-import { Button } from "@/components/ui/buttons"
+import { Button } from "@/components/ui/button"
 
 export function ThemeToggle() {
   const { setTheme, theme } = useTheme()
